// src/interfaces/CommentsInterfaces.tsx
// Generated by https://quicktype.io

export interface CommentsReponse {
  data:  Comments[];
  total: number;
  page:  number;
  limit: number;
}

export interface Comments {
  id:          string;
  message:     string;
  owner:       Owner;
  post:        string;
  publishDate: string;
}

export interface Owner {
  id:        string;
  title:     string;
  firstName: string;
  lastName:  string;
  picture:   string;
}